+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG for Sulu
 * dev-master
     * HOTFIX      #3647 [WebsiteBundle]           Fixed sulu_content_load throwing exception when reference was deleted
     * HOTFIX      #3646 [ContentBundle]           Fixed xss attack for seo description
+    * HOTFIX      #3652 [RouteBundle]             Fixed bug with json routes for route-bundle
 
 * 1.6.8 (2017-11-21)
     * BUGFIX      #3629 [WebsiteBundle]           Homepage route not matched